+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.2.0
 
+Bug fixes:
+
+- No longer hangs when decoding the string "---" (workaround for bug in yamerl)
+
 Notes:
 
 - Now that 'yamerl' has been [published to Hex.pm](https://hex.pm/packages/yamerl), switched from 'yamerl_the_fork' to 'yamerl' for parsing